Fix attempt to load Config class before autoloader
